rename CooperationActionModal to CooperationActionBanner
