+++ src/containers/my-cooperations/cooperation-action-banner/CooperationActionBanner.styles.ts
